pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "naviagent"
version = "0.1.0"
description = "NaviAgent - A navigation agent framework"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "NaviAgent Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "agno==2.1.6",
    "autoflake>=2.0.0",
    "bandit[toml]>=1.7.0",
    "black>=23.0.0",
    "ddgs>=9.6.1",
    "fastapi==0.115.0",
    "flake8>=6.0.0",
    "flake8-bugbear>=23.0.0",
    "flake8-comprehensions>=3.14.0",
    "flake8-docstrings>=1.7.0",
    "flake8-simplify>=0.21.0",
    "googlesearch-python==1.2.5",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "openai>=1.60.0",
    "pre-commit>=3.0.0",
    "pycountry==24.6.1",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "pydocstyle>=6.3.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "uvicorn[standard]==0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "autoflake>=2.0.0",
    "pre-commit>=3.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
format = "naviagent.scripts:format_code"
lint = "naviagent.scripts:lint_code"
check = "naviagent.scripts:check_all"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*_test.py",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = [
    "tests",
    ".venv",
    "venv",
    "build",
    "dist",
]
skips = ["B101", "B601"]

[dependency-groups]
dev = [
  "black>=23.0.0",
  "isort>=5.12.0",
  "flake8>=6.0.0",
  "mypy>=1.0.0",
  "pre-commit>=3.0.0",
]

[tool.uv.workspace]
members = []
